requirements.txt
llama-index
llama-index-llms-openai
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
llama-index-tools-google
python-dotenv
chromadb
transformers
torch
llama-index-tools-mcp